backend/requirements.txt
streamlit
pytesseract
Pillow
pymupdf
qdrant-client
sentence-transformers
openai
llama-index
tika